Login and register system
The first function is mainly to register the login page with index.html, then render it with css, then introduce the required module with sever. Create a connection to MySQL database, specify the host, user name, password and database name, and access this index page with get method. Finally write monitor stop port, so that the local can run. By haoxuanyu