Top 10
This function is to display the top10 pages of the account binding, which will be given to different accounts according to the game you bought, etc. There will be different top10 pages of the account, and I will complete the code in descending order after obtaining the data. By haoxuanyu